% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-functions.R
\name{clean_na_cols}
\alias{clean_na_cols}
\title{Clean NAs in columns}
\usage{
clean_na_cols(X, clean_mode = c("remove", "impute_mean"), verbose = 0)
}
\arguments{
\item{X}{A data frame.}

\item{clean_mode}{A character string. One of "remove" or "impute_mean".}

\item{verbose}{An integer. The level of verbosity.}
}
\value{
A data frame, cleaned without NAs.
}
\description{
This function cleans NAs in columns of a data frame.
}
\examples{
X <- data.frame(a = c(1, 2, NA), b = c(NA, 3, 4), c = c(1, 2, 3))
X_cleaned <- clean_na_cols(X, clean_mode = "remove")
X_cleaned <- clean_na_cols(X, clean_mode = "impute_mean")

}
